Refactor types into separate target
