+++ BUILD.bazel
@@ -352,25 +352,27 @@ gentbl_cc_library(
 )
 
 cc_library(
-    name = "vhlo_ops",
-    srcs = [
-        "stablehlo/dialect/VhloBytecode.cpp",
-        "stablehlo/dialect/VhloOps.cpp",
-    ],
+    name = "vhlo_dialect",
+    srcs = [],
     hdrs = [
-        "stablehlo/dialect/VhloBytecode.h",
-        "stablehlo/dialect/VhloOps.h",
-        "stablehlo/transforms/VhloBuiltinTypeConversion.h",
+        "stablehlo/dialect/VhloDialect.h",
     ],
     deps = [
-        ":base",
-        ":stablehlo_assembly_format",
+        "@llvm-project//mlir:IR",
+    ],
+)
+
+cc_library(
+    name = "vhlo_types",
+    srcs = [
+        "stablehlo/dialect/VhloTypes.cpp",
+    ],
+    hdrs = [
+        "stablehlo/dialect/VhloTypes.h",
+    ],
+    deps = [
         ":version",
-        ":vhlo_attr_interfaces_inc_gen",
-        ":vhlo_attrs_inc_gen",
-        ":vhlo_enums_inc_gen",
-        ":vhlo_op_interfaces_inc_gen",
-        ":vhlo_ops_inc_gen",
+        ":vhlo_dialect",
         ":vhlo_type_interfaces_inc_gen",
         ":vhlo_types_inc_gen",
         "@llvm-project//llvm:Support",
@@ -379,6 +381,35 @@ cc_library(
         "@llvm-project//mlir:ShapeDialect",
         "@llvm-project//mlir:Support",
         "@llvm-project//mlir:Transforms",
+    ],
+)
+
+cc_library(
+    name = "vhlo_ops",
+    srcs = [
+        "stablehlo/dialect/VhloBytecode.cpp",
+        "stablehlo/dialect/VhloOps.cpp",
+    ],
+    hdrs = [
+        "stablehlo/dialect/VhloBytecode.h",
+        "stablehlo/dialect/VhloOps.h",
+    ],
+    deps = [
+        ":base",
+        ":stablehlo_assembly_format",
+        ":version",
+        ":vhlo_attr_interfaces_inc_gen",
+        ":vhlo_attrs_inc_gen",
+        ":vhlo_dialect",
+        ":vhlo_enums_inc_gen",
+        ":vhlo_op_interfaces_inc_gen",
+        ":vhlo_ops_inc_gen",
+        ":vhlo_types",
+        "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:QuantOps",
+        "@llvm-project//mlir:ShapeDialect",
+        "@llvm-project//mlir:Support",
     ],
 )
 
@@ -493,6 +524,7 @@ cc_library(
         ":stablehlo_type_inference",
         ":version",
         ":vhlo_ops",
+        ":vhlo_types",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:FuncDialect",
         "@llvm-project//mlir:InferTypeOpInterface",
